fit and predict method completed

+++ src/glow/tensor_numpy_adapter.py
@@ -7,8 +7,9 @@ class TensorNumpyAdapter:
     type and Tensor objects in PyTorch
 
     """
+
     def to_tensor(self, x):
-        return torch.from_numpy(x)
+        return torch.from_numpy(x).float()
 
     def to_numpy(self, x):
         return x.numpy()
